dgbmv on Poisson1D and EX_SOL
dgbtrftridiag : small fix, replace kl by kv for the zero column

.mk file : adding new library

In order to validate the storage of the Poisson1D matrix,
we use the dgbmv function to multiply the POISSON matrix with
the EX_SOL vector. If we get a result very close to RHS,
the storage can be considered as correct.

however, despite what the dgbmv documentation seems to say,
the zero (kv) buffer column is not on the first column but on the last.
To adjust, we pass the matrix AB+1 as an argument.

Finally, I don't know if the problem comes from the storage
of the AB matrix or from the dgbmv function.